knight implementation tweaks
